Reduce memory usage when compiling oneflow dialect ops (#7000)
