+++ oneflow/ir/lib/OneFlow/OneFlowOpGetGen.cpp.in
@@ -1,7 +1,5 @@
-#include "OneFlow/OneFlowOps.h"
 #include <iostream>
 #include <string>
-#include "OneFlow/OneFlowDialect.h"
 #include "OneFlow/Passes.h"
 #include "llvm/ADT/STLExtras.h"
 #include "mlir/IR/BuiltinAttributes.h"
@@ -10,6 +8,13 @@
 #include "llvm/ADT/StringSet.h"
 #include "mlir/Support/LLVM.h"
 #include "mlir/Support/LogicalResult.h"
+#include "OneFlow/OneFlowDialect.h"
+#include "OneFlow/OneFlowOpTraits.h"
+#include "OneFlow/OneFlowEnums.h.inc"
+#include "OneFlow/OneFlowSupport.h"
+#include "OneFlow/OneFlowInterfaces.h.inc"
 
+#define GET_OP_CLASSES
+#include "OneFlow/OneFlow.@OP_GROUP_NAME_LOWER@_ops.h.inc"
 #define GET_OP_CLASSES
 #include "OneFlow/OneFlow.@OP_GROUP_NAME_LOWER@_ops.cpp.inc"
